Recursive expansion of tuple
The original code assumes only the right expression can be a tuple, i.e. all tuples are of the form (A * (B * (C * D))). But users are able to construct tuples of the form ((A * B) * (C * D)). In such cases the original code gives the unexpected expanded list of [(A * B), C, D]. It's hard to support multi-level tuple structure, so we should recursively transform it to a completely flatten list [A, B, C, D] (instead of [A * B, C * D]).